Updated to later Raze version
Got stereo separation working correctly

+++ Projects/Android/jni/Raze/source/common/utility/cmdlib.cpp
@@ -248,7 +248,7 @@ bool GetFileInfo(const char* pathname, size_t *size, time_t *time)
 	bool res = _wstat64(wstr.c_str(), &info) == 0;
 #endif
 	if (!res || (info.st_mode & S_IFDIR)) return false;
-	if (size) *size = info.st_size;
+	if (size) *size = (size_t)info.st_size;
 	if (time) *time = info.st_mtime;
 	return res;
 }
@@ -557,7 +557,7 @@ void CreatePath(const char *fn)
 void CreatePath(const char *fn)
 {
 	char *copy, *p;
- 
+
 	if (fn[0] == '/' && fn[1] == '\0')
 	{
 		return;
