+++ Projects/Android/jni/Raze/source/games/exhumed/src/menu.cpp
@@ -16,7 +16,6 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 */
 //-------------------------------------------------------------------------
 #include "ns.h"
-#include "compat.h"
 #include "build.h"
 #include "exhumed.h"
 #include "aistuff.h"
@@ -51,7 +50,7 @@ uint8_t *dest;
 
 unsigned int nRandom = 0x41C6167E;
 int dword_9AB57 = 0x1F;
-short word_9AB5B = 0;
+int word_9AB5B = 0;
 
 int keytimer = 0;
 
@@ -87,7 +86,7 @@ void DoEnergyTile()
     uint8_t* ptr1 = energy1 + 1984;
     uint8_t* ptr2 = energy1 + 2048;
 
-    short nColor = nButtonColor + 161;
+    int nColor = nButtonColor + 161;
 
     int i, j;
